Cloud functions code/Analyzing and Tagging/requirements.txt
azure-storage-blob
azure-cosmos
requests
msrest
azure-identity
azure-cognitiveservices-vision-customvision